code cleanup
git-svn-id: svn+ssh://svn.berlios.de/svnroot/repos/awds/routing@205 d2c7b242-0724-0410-acbd-db3348fdb001

+++ src/AwdsRouting.cc
@@ -564,13 +564,13 @@ void awds::AwdsRouting::recv_unicast(BasePacket *p) {
     bool found;
     topology->getNextHop(dest, nextHop, found);
     if (!found) {
-	GEA.dbg() << "no route to host " << (int)dest << std::endl;
+	GEA.dbg() << "no route to host " << dest << std::endl;
 	//	topology->print();
 	return;
     } else {
 	if (verbose) {
-	    GEA.dbg() << "next hop to " << (int) dest 
-		      << " is " << (int) nextHop 
+	    GEA.dbg() << "next hop to " << dest 
+		      << " is " << nextHop 
 		      << std::endl;
 	}
